Add integration tests helper on test package
